% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dye.R
\name{dyes}
\alias{dyes}
\alias{dyes<-}
\title{Dyes for a DNA mixture}
\usage{
dyes(mixture)

dyes(mixture) <- value
}
\arguments{
\item{mixture}{A DNA mixture}

\item{value}{A list with one item per dye, each containing a vector of marker names.}
}
\value{
The dyes used for the DNA mixture
}
\description{
\emph{IMPORTANT: This is the \pkg{DNAmixturesLite}
    package, which is intended as a service to enable users to try
    \pkg{DNAmixtures} without purchasing a commercial licence for
    Hugin. When at all possible, we strongly recommend the use of
    \pkg{DNAmixtures} rather than this lite-version. See
    \url{https://dnamixtures.r-forge.r-project.org/} for details
    on both packages.}

\emph{While the lite-version seeks to provide the full functionality of
\pkg{DNAmixtures}, note that computations are much less efficient
and that there are some differences in available functionality. Be
aware that the present documentation is copied from
\pkg{DNAmixtures} and thus may not accurately describe the
implementation of this lite-version.}
}
\seealso{
\code{\link{DNAmixture}}
}
