% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDescriptions.R
\docType{data}
\name{MC15}
\alias{MC15}
\title{The MC15 DNA mixture}
\format{
A \code{data.frame}.
}
\source{
Peter Gill et. al. (2008)
\emph{Interpretation of complex {DNA} profiles using empirical models and a method to measure their robustness}.
Forensic Science International: Genetics, 2(2):91--103.
}
\description{
Peak heights from analysis of a bloodstain with DNA
from an unknown number of contributors. There are three reference
profiles \code{K1}, \code{K2}, and \code{K3} for individuals
related to the case.
}
