% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDescriptions.R
\docType{data}
\name{NGM}
\alias{NGM}
\title{NGM allele frequencies}
\format{
A list containing 
\describe{
\item{\code{USCaucasian}}{A \code{data.frame} with, for each marker,
the set of possible alleles as well as their corresponding frequency.}
}
}
\source{
Budowle et. al (2011) \emph{Population Genetic Analyses of the NGM STR loci}.
International Journal of Legal Medicine, 125(1):101-109.
}
\description{
NGM allele frequencies.
}
