% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDescriptions.R
\docType{data}
\name{NGMDyes}
\alias{NGMDyes}
\title{Dyes used for NGM}
\format{
A list with one item per dye, each containing a vector of
marker-names specifying the order in which they occur in an EPG.
}
\source{
Applied Biosystems
}
\description{
Dyes used for the AmpFlSTR NGM PCR Amplification Kit
}
