% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDescriptions.R
\docType{data}
\name{SGMplusDyes}
\alias{SGMplusDyes}
\title{Dyes used for SGMplus}
\format{
A list with one item per dye, each containing a vector of
marker-names specifying the order in which they occur in an EPG.
}
\source{
Applied Biosystems
}
\description{
Dyes used for the AmpFlSTR SGM Plus PCR Amplification Kit
}
