% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDescriptions.R
\docType{data}
\name{UKCaucasian}
\alias{UKCaucasian}
\title{Allele frequencies for UK Caucasians}
\format{
A \code{data.frame}
}
\source{
The frequencies are produced from the allele counts for
one (the UK Caucasian) of three British sub-populations as found
on David Balding's homepage under his
\href{https://sites.google.com/site/baldingstatisticalgenetics/software/likeltd-r-forensic-dna-r-code}{\pkg{likeLTD}}
software for mixture analysis in \R.
}
\description{
Database of allele frequencies for UK Caucasians.
}
\references{
\url{https://sites.google.com/site/baldingstatisticalgenetics/software/likeltd-r-forensic-dna-r-code}
}
