% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDescriptions.R
\docType{data}
\name{USCaucasian}
\alias{USCaucasian}
\title{The data base of allele frequencies for 302 US Caucasian profiles.}
\format{
A \code{data.frame}.
}
\source{
\url{http://www.cstl.nist.gov/strbase/NISTpop.htm}
}
\description{
Allele frequencies for US Caucasians.
}
\note{
There are a few errata found after publication of the allele
frequencies, but we have not updated the allele frequencies
accordingly. The frequencies also did not add up to 1, and this is
simply corrected by normalising within each marker.
}
